refactor(test): share hook-recording helper, clear stale render-cache text
Extract the React use*-recording harness (duplicated across the web hook
tests and the new native one) into src/test/recordHooks.ts, and update the
tests, docs, and agent rules that still described the removed render cache:
retitle the HMR class-name test, rewrite the memoization suite's header and
hook-sequence test to re-render vocabulary, correct the runtime-performance
hot-path wording, and fix an AGENTS rule that read as "hooks must be
unconditional" against the deliberate __SERVER__/IS_RSC gating. Trim the
changeset to user-observable behavior.
