+++ packages/styled-components/src/native/test/native.test.tsx
@@ -3,6 +3,7 @@ import { Image, Text, View, ViewProps } from 'react-native';
 import TestRenderer from 'react-test-renderer';
 import styled, { ThemeProvider, css, toStyleSheet } from '../';
 import { RN_UNSUPPORTED_VALUES } from '../../models/InlineStyle';
+import { withHookRecording } from '../../test/recordHooks';
 
 // NOTE: These tests are like the ones for Web but a "light-version" of them
 // This is mostly due to the similar logic
@@ -595,40 +596,14 @@ describe('native', () => {
     });
   });
 
-  // Mirrors the web coverage in src/test/memoization.test.tsx: skipping style
-  // evaluation on an unchanged-props re-render would drop a hook called inside an
-  // interpolation and serve a stale style object (#5788).
+  // Mirrors the web coverage in src/test/memoization.test.tsx: an interpolation,
+  // and any hook it calls or external value it reads, runs on every render and is
+  // not skipped on an unchanged-props re-render (#5788).
   describe('interpolation evaluation on every render (#5788)', () => {
     it('runs a hook called inside an interpolation on every render', () => {
-      const reactHooks = React as unknown as Record<string, (...args: unknown[]) => unknown>;
       const Ctx = React.createContext('0.5');
-      const hookNames = Object.keys(React).filter(
-        key => key.startsWith('use') && typeof reactHooks[key] === 'function'
-      );
 
-      let recording: string[] | null = null;
-      const originals = new Map<string, (...args: unknown[]) => unknown>();
-
-      function record<T>(render: () => T): [T, string[]] {
-        const calls: string[] = [];
-        recording = calls;
-        try {
-          return [render(), calls];
-        } finally {
-          recording = null;
-        }
-      }
-
-      try {
-        for (const name of hookNames) {
-          const original = reactHooks[name];
-          originals.set(name, original);
-          reactHooks[name] = (...args) => {
-            if (recording) recording.push(name);
-            return original(...args);
-          };
-        }
-
+      withHookRecording(record => {
         const Comp = styled.View<{ $pad: number }>`
           padding: ${p => p.$pad}px;
           opacity: ${() => React.useContext(Ctx)};
@@ -641,7 +616,8 @@ describe('native', () => {
             </Ctx.Provider>
           )
         );
-        const [, cacheHit] = record(() =>
+        // Same props: the interpolation, and the useContext it calls, still run.
+        const [, reRender] = record(() =>
           wrapper.update(
             <Ctx.Provider value="0.5">
               <Comp $pad={1} />
@@ -650,12 +626,10 @@ describe('native', () => {
         );
 
         expect(mount).toContain('useContext');
-        expect(cacheHit).toEqual(mount);
+        expect(reRender).toEqual(mount);
 
         wrapper.unmount();
-      } finally {
-        for (const [name, original] of originals) reactHooks[name] = original;
-      }
+      });
     });
 
     it('recomputes when an interpolation reads external state that changed but props did not', () => {
@@ -670,7 +644,7 @@ describe('native', () => {
       expect(wrapper.root.findByType(View).props.style).toEqual({ color: 'red', ...padding });
 
       external = 'blue';
-      // Same props: the former cache would hit and return the stale red style.
+      // Same props, but the interpolation's external input changed.
       wrapper.update(<Comp $pad={1} />);
       expect(wrapper.root.findByType(View).props.style).toEqual({ color: 'blue', ...padding });
